refactor(shared): 归位存储布局契约
