+++ docs/specs/storage-layout-spec.md
@@ -1,6 +1,6 @@
 # OMK storage layout v2
 
-> **Scope:** the canonical project and machine storage contract. Paths are generated by `src/omk-layout/`; product modules must not assemble `.omk` roots themselves. This layout changes storage only and does not change report schemas, scoring, prompts, statistics, or length debiasing.
+> **Scope:** the canonical project and machine storage contract. Paths are generated by `src/shared/storage-layout.ts`; product modules must not assemble `.omk` roots themselves. This layout changes storage only and does not change report schemas, scoring, prompts, statistics, or length debiasing.
 
 ## Principles
