+++ docs/zh/specs/storage-layout-spec.md
@@ -1,6 +1,6 @@
 # OMK 存储布局 v2
 
-> **范围：**项目级与机器级存储的正式契约。所有路径由 `src/omk-layout/` 统一生成，业务模块不得自行拼项目根 `.omk`。本次只改变存储位置，不改变报告 schema、评分语义、prompt、统计公式或长度去偏。
+> **范围：**项目级与机器级存储的正式契约。所有路径由 `src/shared/storage-layout.ts` 统一生成，业务模块不得自行拼项目根 `.omk`。本次只改变存储位置，不改变报告 schema、评分语义、prompt、统计公式或长度去偏。
 
 ## 设计原则
