+++ src/cli/commands/init.ts
@@ -4,7 +4,7 @@ import { LANG_FLAG, bilingual, resolveLang } from '../oclif/i18n.js';
 import { BaseCommand } from '../oclif/base-command.js';
 import { tCli } from '../lib/i18n.js';
 import { shellQuoteArg } from '../../shared/shell-quote.js';
-import { projectLayout } from '../../omk-layout/index.js';
+import { projectLayout } from '../../shared/storage-layout.js';
 import {
   DEFAULT_INIT_SAMPLE_COUNT,
   FULL_INIT_SAMPLE_COUNT,
