+++ src/shared/storage-layout.ts
